Struct aws_sdk_lexmodelsv2::client::fluent_builders::UpdateIntent

pub struct UpdateIntent<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateIntent<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateIntentOutput, SdkError<UpdateIntentError>> where
R::Policy: SmithyRetryPolicy<UpdateIntentInputOperationOutputAlias, UpdateIntentOutput, UpdateIntentError, UpdateIntentInputOperationRetryAlias>,

pub fn intent_id(self, inp: impl Into<String>) -> Self

The unique identifier of the intent to update.

pub fn set_intent_id(self, input: Option<String>) -> Self

pub fn intent_name(self, inp: impl Into<String>) -> Self

The new name for the intent.

pub fn set_intent_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The new description of the intent.

pub fn set_description(self, input: Option<String>) -> Self

pub fn parent_intent_signature(self, inp: impl Into<String>) -> Self

The signature of the new built-in intent to use as the parent of this intent.

pub fn set_parent_intent_signature(self, input: Option<String>) -> Self

pub fn sample_utterances(self, inp: impl Into<SampleUtterance>) -> Self

Appends an item to sampleUtterances.

To override the contents of this collection use set_sample_utterances.

New utterances used to invoke the intent.

pub fn set_sample_utterances(self, input: Option<Vec<SampleUtterance>>) -> Self

pub fn dialog_code_hook(self, inp: DialogCodeHookSettings) -> Self

The new Lambda function to use between each turn of the conversation with the bot.

pub fn set_dialog_code_hook(self, input: Option<DialogCodeHookSettings>) -> Self

pub fn fulfillment_code_hook(self, inp: FulfillmentCodeHookSettings) -> Self

The new Lambda function to call when all of the intents required slots are provided and the intent is ready for fulfillment.

pub fn set_fulfillment_code_hook(
self,
input: Option<FulfillmentCodeHookSettings>
) -> Self

pub fn slot_priorities(self, inp: impl Into<SlotPriority>) -> Self

Appends an item to slotPriorities.

To override the contents of this collection use set_slot_priorities.

A new list of slots and their priorities that are contained by the intent.

pub fn set_slot_priorities(self, input: Option<Vec<SlotPriority>>) -> Self

pub fn intent_confirmation_setting(self, inp: IntentConfirmationSetting) -> Self

New prompts that Amazon Lex sends to the user to confirm the completion of an intent.

pub fn set_intent_confirmation_setting(
self,
input: Option<IntentConfirmationSetting>
) -> Self

pub fn intent_closing_setting(self, inp: IntentClosingSetting) -> Self

The new response that Amazon Lex sends the user when the intent is closed.

pub fn set_intent_closing_setting(
self,
input: Option<IntentClosingSetting>
) -> Self

pub fn input_contexts(self, inp: impl Into<InputContext>) -> Self

Appends an item to inputContexts.

To override the contents of this collection use set_input_contexts.

A new list of contexts that must be active in order for Amazon Lex to consider the intent.

pub fn set_input_contexts(self, input: Option<Vec<InputContext>>) -> Self

pub fn output_contexts(self, inp: impl Into<OutputContext>) -> Self

Appends an item to outputContexts.

To override the contents of this collection use set_output_contexts.

A new list of contexts that Amazon Lex activates when the intent is fulfilled.

pub fn set_output_contexts(self, input: Option<Vec<OutputContext>>) -> Self

pub fn kendra_configuration(self, inp: KendraConfiguration) -> Self

New configuration settings for connecting to an Amazon Kendra index.

pub fn set_kendra_configuration(
self,
input: Option<KendraConfiguration>
) -> Self

pub fn bot_id(self, inp: impl Into<String>) -> Self

The identifier of the bot that contains the intent.

pub fn set_bot_id(self, input: Option<String>) -> Self

pub fn bot_version(self, inp: impl Into<String>) -> Self

The version of the bot that contains the intent. Must be DRAFT.

pub fn set_bot_version(self, input: Option<String>) -> Self

pub fn locale_id(self, inp: impl Into<String>) -> Self

The identifier of the language and locale where this intent is used. The string must match one of the supported locales. For more information, see Supported languages.

pub fn set_locale_id(self, input: Option<String>) -> Self