Struct aws_sdk_lexmodelsv2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for LexModelBuildingServiceV2.

This client allows ergonomic access to a LexModelBuildingServiceV2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn build_bot_locale(&self) -> BuildBotLocale<C, M, R>

pub fn create_bot(&self) -> CreateBot<C, M, R>

pub fn create_bot_alias(&self) -> CreateBotAlias<C, M, R>

pub fn create_bot_locale(&self) -> CreateBotLocale<C, M, R>

pub fn create_bot_version(&self) -> CreateBotVersion<C, M, R>

pub fn create_export(&self) -> CreateExport<C, M, R>

pub fn create_intent(&self) -> CreateIntent<C, M, R>

pub fn create_resource_policy(&self) -> CreateResourcePolicy<C, M, R>

pub fn create_resource_policy_statement(
&self
) -> CreateResourcePolicyStatement<C, M, R>

pub fn create_slot(&self) -> CreateSlot<C, M, R>

pub fn create_slot_type(&self) -> CreateSlotType<C, M, R>

pub fn create_upload_url(&self) -> CreateUploadUrl<C, M, R>

pub fn delete_bot(&self) -> DeleteBot<C, M, R>

pub fn delete_bot_alias(&self) -> DeleteBotAlias<C, M, R>

pub fn delete_bot_locale(&self) -> DeleteBotLocale<C, M, R>

pub fn delete_bot_version(&self) -> DeleteBotVersion<C, M, R>

pub fn delete_export(&self) -> DeleteExport<C, M, R>

pub fn delete_import(&self) -> DeleteImport<C, M, R>

pub fn delete_intent(&self) -> DeleteIntent<C, M, R>

pub fn delete_resource_policy(&self) -> DeleteResourcePolicy<C, M, R>

pub fn delete_resource_policy_statement(
&self
) -> DeleteResourcePolicyStatement<C, M, R>

pub fn delete_slot(&self) -> DeleteSlot<C, M, R>

pub fn delete_slot_type(&self) -> DeleteSlotType<C, M, R>

pub fn describe_bot(&self) -> DescribeBot<C, M, R>

pub fn describe_bot_alias(&self) -> DescribeBotAlias<C, M, R>

pub fn describe_bot_locale(&self) -> DescribeBotLocale<C, M, R>

pub fn describe_bot_version(&self) -> DescribeBotVersion<C, M, R>

pub fn describe_export(&self) -> DescribeExport<C, M, R>

pub fn describe_import(&self) -> DescribeImport<C, M, R>

pub fn describe_intent(&self) -> DescribeIntent<C, M, R>

pub fn describe_resource_policy(&self) -> DescribeResourcePolicy<C, M, R>

pub fn describe_slot(&self) -> DescribeSlot<C, M, R>

pub fn describe_slot_type(&self) -> DescribeSlotType<C, M, R>

pub fn list_bot_aliases(&self) -> ListBotAliases<C, M, R>

pub fn list_bot_locales(&self) -> ListBotLocales<C, M, R>

pub fn list_bots(&self) -> ListBots<C, M, R>

pub fn list_bot_versions(&self) -> ListBotVersions<C, M, R>

pub fn list_built_in_intents(&self) -> ListBuiltInIntents<C, M, R>

pub fn list_built_in_slot_types(&self) -> ListBuiltInSlotTypes<C, M, R>

pub fn list_exports(&self) -> ListExports<C, M, R>

pub fn list_imports(&self) -> ListImports<C, M, R>

pub fn list_intents(&self) -> ListIntents<C, M, R>

pub fn list_slots(&self) -> ListSlots<C, M, R>

pub fn list_slot_types(&self) -> ListSlotTypes<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn start_import(&self) -> StartImport<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_bot(&self) -> UpdateBot<C, M, R>

pub fn update_bot_alias(&self) -> UpdateBotAlias<C, M, R>

pub fn update_bot_locale(&self) -> UpdateBotLocale<C, M, R>

pub fn update_export(&self) -> UpdateExport<C, M, R>

pub fn update_intent(&self) -> UpdateIntent<C, M, R>

pub fn update_resource_policy(&self) -> UpdateResourcePolicy<C, M, R>

pub fn update_slot(&self) -> UpdateSlot<C, M, R>

pub fn update_slot_type(&self) -> UpdateSlotType<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self