Module aws_sdk_lexmodelsv2::model

Modules

audio_log_destination

See AudioLogDestination

audio_log_setting

See AudioLogSetting

bot_alias_history_event

See BotAliasHistoryEvent

bot_alias_locale_settings

See BotAliasLocaleSettings

bot_alias_summary

See BotAliasSummary

bot_export_specification

See BotExportSpecification

bot_filter

See BotFilter

bot_import_specification

See BotImportSpecification

bot_locale_export_specification

See BotLocaleExportSpecification

bot_locale_filter

See BotLocaleFilter

bot_locale_history_event

See BotLocaleHistoryEvent

bot_locale_import_specification

See BotLocaleImportSpecification

bot_locale_sort_by

See BotLocaleSortBy

bot_locale_summary

See BotLocaleSummary

bot_sort_by

See BotSortBy

bot_summary

See BotSummary

bot_version_locale_details

See BotVersionLocaleDetails

bot_version_sort_by

See BotVersionSortBy

bot_version_summary

See BotVersionSummary

built_in_intent_sort_by

See BuiltInIntentSortBy

built_in_intent_summary

See BuiltInIntentSummary

built_in_slot_type_sort_by

See BuiltInSlotTypeSortBy

built_in_slot_type_summary

See BuiltInSlotTypeSummary

button

See Button

cloud_watch_log_group_log_destination

See CloudWatchLogGroupLogDestination

code_hook_specification

See CodeHookSpecification

conversation_log_settings

See ConversationLogSettings

custom_payload

See CustomPayload

data_privacy

See DataPrivacy

dialog_code_hook_settings

See DialogCodeHookSettings

export_filter

See ExportFilter

export_resource_specification

See ExportResourceSpecification

export_sort_by

See ExportSortBy

export_summary

See ExportSummary

fulfillment_code_hook_settings

See FulfillmentCodeHookSettings

image_response_card

See ImageResponseCard

import_filter

See ImportFilter

import_resource_specification

See ImportResourceSpecification

import_sort_by

See ImportSortBy

import_summary

See ImportSummary

input_context

See InputContext

intent_closing_setting

See IntentClosingSetting

intent_confirmation_setting

See IntentConfirmationSetting

intent_filter

See IntentFilter

intent_sort_by

See IntentSortBy

intent_summary

See IntentSummary

kendra_configuration

See KendraConfiguration

lambda_code_hook

See LambdaCodeHook

message

See Message

message_group

See MessageGroup

multiple_values_setting

See MultipleValuesSetting

obfuscation_setting

See ObfuscationSetting

output_context

See OutputContext

plain_text_message

See PlainTextMessage

principal

See Principal

prompt_specification

See PromptSpecification

response_specification

See ResponseSpecification

s3_bucket_log_destination

See S3BucketLogDestination

sample_utterance

See SampleUtterance

sample_value

See SampleValue

sentiment_analysis_settings

See SentimentAnalysisSettings

slot_default_value

See SlotDefaultValue

slot_default_value_specification

See SlotDefaultValueSpecification

slot_filter

See SlotFilter

slot_priority

See SlotPriority

slot_sort_by

See SlotSortBy

slot_summary

See SlotSummary

slot_type_filter

See SlotTypeFilter

slot_type_sort_by

See SlotTypeSortBy

slot_type_summary

See SlotTypeSummary

slot_type_value

See SlotTypeValue

slot_value_elicitation_setting

See SlotValueElicitationSetting

slot_value_regex_filter

See SlotValueRegexFilter

slot_value_selection_setting

See SlotValueSelectionSetting

ssml_message

See SsmlMessage

still_waiting_response_specification

See StillWaitingResponseSpecification

text_log_destination

See TextLogDestination

text_log_setting

See TextLogSetting

voice_settings

See VoiceSettings

wait_and_continue_specification

See WaitAndContinueSpecification

Structs

AudioLogDestination

The location of audio log files collected when conversation logging

AudioLogSetting

Settings for logging audio of conversations between Amazon Lex and a

BotAliasHistoryEvent

Provides a record of an event that affects a bot alias. For example,

BotAliasLocaleSettings

Specifies settings that are unique to a locale. For example, you can

BotAliasSummary

Summary information about bot aliases returned from the ListBotAliases operation.

BotExportSpecification

Provides the identity of a the bot that was exported.

BotFilter

Filters the responses returned by the ListBots

BotImportSpecification

Provides the bot parameters required for importing a bot.

BotLocaleExportSpecification

Provides the bot locale parameters required for exporting a bot

BotLocaleFilter

Filters responses returned by the ListBotLocales

BotLocaleHistoryEvent

Provides information about an event that occurred affecting the bot

BotLocaleImportSpecification

Provides the bot locale parameters required for importing a bot

BotLocaleSortBy

Specifies attributes for sorting a list of bot locales.

BotLocaleSummary

Summary information about bot locales returned by the ListBotLocales operation.

BotSortBy

Specifies attributes for sorting a list of bots.

BotSummary

Summary information about a bot returned by the ListBots operation.

BotVersionLocaleDetails

The version of a bot used for a bot locale.

BotVersionSortBy

Specifies attributes for sorting a list of bot versions.

BotVersionSummary

Summary information about a bot version returned by the ListBotVersions operation.

BuiltInIntentSortBy

Specifies attributes for sorting a list of built-in intents.

BuiltInIntentSummary

Provides summary information about a built-in intent for the ListBuiltInIntents operation.

BuiltInSlotTypeSortBy

Specifies attributes for sorting a list of built-in slot

BuiltInSlotTypeSummary

Provides summary information about a built-in slot type for the

Button

Describes a button to use on a response card used to gather slot

CloudWatchLogGroupLogDestination

The Amazon CloudWatch Logs log group where the text and metadata logs are

CodeHookSpecification

Contains information about code hooks that Amazon Lex calls during a

ConversationLogSettings

Configures conversation logging that saves audio, text, and metadata

CustomPayload

A custom response string that Amazon Lex sends to your application. You

DataPrivacy

By default, data stored by Amazon Lex is encrypted. The

DialogCodeHookSettings

Settings that determine the Lambda function that Amazon Lex uses for

ExportFilter

Filtes the response form the

ExportResourceSpecification

Provides information about the bot or bot locale that you want to

ExportSortBy

Provides information about sorting a list of exports.

ExportSummary

Provides summary information about an export in an export list.

FulfillmentCodeHookSettings

Determines if a Lambda function should be invoked for a specific

ImageResponseCard

A card that is shown to the user by a messaging platform. You define

ImportFilter

Filters the response from the

ImportResourceSpecification

Provides information about the bot or bot locale that you want to

ImportSortBy

Provides information for sorting a list of imports.

ImportSummary

Provides summary information about an import in an import

InputContext

The name of a context that must be active for an intent to be

IntentClosingSetting

Provides a statement the Amazon Lex conveys to the user when the intent

IntentConfirmationSetting

Provides a prompt for making sure that the user is ready for the

IntentFilter

Filters the response from the ListIntents

IntentSortBy

Specifies attributes for sorting a list of intents.

IntentSummary

Summary information about an intent returned by the

KendraConfiguration

Provides configuration information for the AMAZON.KendraSearchIntent

LambdaCodeHook

Specifies a Lambda function that verifies requests to a bot or

Message

The object that provides message text and it's type.

MessageGroup

Provides one or more messages that Amazon Lex should send to the

MultipleValuesSetting

Indicates whether a slot can return multiple values.

ObfuscationSetting

Determines whether Amazon Lex obscures slot values in conversation logs.

OutputContext

Describes a session context that is activated when an intent is

PlainTextMessage

Defines an ASCII text message to send to the user.

Principal

The IAM principal that you allowing or denying access to an Amazon Lex

PromptSpecification

Specifies a list of message groups that Amazon Lex sends to a user to

ResponseSpecification

Specifies a list of message groups that Amazon Lex uses to respond the

S3BucketLogDestination

Specifies an Amazon S3 bucket for logging audio conversations

SampleUtterance

A sample utterance that invokes an intent or respond to a slot

SampleValue

Defines one of the values for a slot type.

SentimentAnalysisSettings

Determines whether Amazon Lex will use Amazon Comprehend to detect the sentiment of

SlotDefaultValue

Specifies the default value to use when a user doesn't provide a

SlotDefaultValueSpecification

Defines a list of values that Amazon Lex should use as the default value

SlotFilter

Filters the response from the ListSlots

SlotPriority

Sets the priority that Amazon Lex should use when eliciting slot values

SlotSortBy

Specifies attributes for sorting a list of bots.

SlotSummary

Summary information about a slot, a value that the bot elicits from

SlotTypeFilter

Filters the response from the ListSlotTypes

SlotTypeSortBy

Specifies attributes for sorting a list of slot types.

SlotTypeSummary

Provides summary information about a slot type.

SlotTypeValue

Each slot type can have a set of values. Each

SlotValueElicitationSetting

Settings that you can use for eliciting a slot value.

SlotValueRegexFilter

Provides a regular expression used to validate the value of a

SlotValueSelectionSetting

Contains settings used by Amazon Lex to select a slot value.

SsmlMessage

Defines a Speech Synthesis Markup Language (SSML) prompt.

StillWaitingResponseSpecification

Defines the messages that Amazon Lex sends to a user to remind them that

TextLogDestination

Defines the Amazon CloudWatch Logs destination log group for

TextLogSetting

Defines settings to enable text conversation logs.

VoiceSettings

Defines settings for using an Amazon Polly voice to communicate with a

WaitAndContinueSpecification

Specifies the prompts that Amazon Lex uses while a bot is waiting for

Enums

BotAliasStatus

BotFilterName

BotFilterOperator

BotLocaleFilterName

BotLocaleFilterOperator

BotLocaleSortAttribute

BotLocaleStatus

BotSortAttribute

BotStatus

BotVersionSortAttribute

BuiltInIntentSortAttribute

BuiltInSlotTypeSortAttribute

Effect

ExportFilterName

ExportFilterOperator

ExportSortAttribute

ExportStatus

ImportExportFileFormat

ImportFilterName

ImportFilterOperator

ImportSortAttribute

ImportStatus

IntentFilterName

IntentFilterOperator

IntentSortAttribute

MergeStrategy

ObfuscationSettingType

SlotConstraint

SlotFilterName

SlotFilterOperator

SlotSortAttribute

SlotTypeFilterName

SlotTypeFilterOperator

SlotTypeSortAttribute

SlotValueResolutionStrategy

SortOrder