Module aws_sdk_lexmodelsv2::operation

Structs

BuildBotLocale

Builds a bot, its intents, and its slot types into a specific

CreateBot

Creates an Amazon Lex conversational bot.

CreateBotAlias

Creates an alias for the specified version of a bot. Use an alias to

CreateBotLocale

Creates a locale in the bot. The locale contains the intents and

CreateBotVersion

Creates a new version of the bot based on the DRAFT

CreateExport

Creates a zip archive containing the contents of a bot or a bot

CreateIntent

Creates an intent.

CreateResourcePolicy

Creates a new resource policy with the specified policy

CreateResourcePolicyStatement

Adds a new resource policy statement to a bot or bot alias. If a

CreateSlot

Creates a slot in an intent. A slot is a variable needed to fulfill

CreateSlotType

Creates a custom slot type

CreateUploadUrl

Gets a pre-signed S3 write URL that you use to upload the zip

DeleteBot

Deletes all versions of a bot, including the Draft

DeleteBotAlias

Deletes the specified bot alias.

DeleteBotLocale

Removes a locale from a bot.

DeleteBotVersion

Deletes a specific version of a bot. To delete all version of a bot,

DeleteExport

Removes a previous export and the associated files stored in an S3

DeleteImport

Removes a previous import and the associated file stored in an S3

DeleteIntent

Removes the specified intent.

DeleteResourcePolicy

Removes an existing policy from a bot or bot alias. If the resource

DeleteResourcePolicyStatement

Deletes a policy statement from a resource policy. If you delete the

DeleteSlot

Deletes the specified slot from an intent.

DeleteSlotType

Deletes a slot type from a bot locale.

DescribeBot

Provides metadata information about a bot.

DescribeBotAlias

Get information about a specific bot alias.

DescribeBotLocale

Describes the settings that a bot has for a specific locale.

DescribeBotVersion

Provides metadata about a version of a bot.

DescribeExport

Gets information about a specific export.

DescribeImport

Gets information about a specific import.

DescribeIntent

Returns metadata about an intent.

DescribeResourcePolicy

Gets the resource policy and policy revision for a bot or bot

DescribeSlot

Gets metadata information about a slot.

DescribeSlotType

Gets metadata information about a slot type.

ListBotAliases

Gets a list of aliases for the specified bot.

ListBotLocales

Gets a list of locales for the specified bot.

ListBotVersions

Gets information about all of the versions of a bot.

ListBots

Gets a list of available bots.

ListBuiltInIntents

Gets a list of built-in intents provided by Amazon Lex that you can use

ListBuiltInSlotTypes

Gets a list of built-in slot types that meet the specified

ListExports

Lists the exports for a bot or bot locale. Exports are kept in the

ListImports

Lists the imports for a bot or bot locale. Imports are kept in the

ListIntents

Get a list of intents that meet the specified criteria.

ListSlotTypes

Gets a list of slot types that match the specified criteria.

ListSlots

Gets a list of slots that match the specified criteria.

ListTagsForResource

Gets a list of tags associated with a resource. Only bots, bot

StartImport

Starts importing a bot or bot locale from a zip archive that you

TagResource

Adds the specified tags to the specified resource. If a tag key

UntagResource

Removes tags from a bot, bot alias, or bot channel.

UpdateBot

Updates the configuration of an existing bot.

UpdateBotAlias

Updates the configuration of an existing bot alias.

UpdateBotLocale

Updates the settings that a bot has for a specific locale.

UpdateExport

Updates the password used to protect an export zip archive.

UpdateIntent

Updates the settings for an intent.

UpdateResourcePolicy

Replaces the existing resource policy for a bot or bot alias with a

UpdateSlot

Updates the settings for a slot.

UpdateSlotType

Updates the configuration of an existing slot type.